requirements.txt
discord.py==2.4.0
dotenv==0.21.0
tabulate==0.9.0